Move toil.batchSystems.MemoryString to toil.lib.conversions.MemoryString.
