+++ src/toil/lib/conversions.py
@@ -3,6 +3,7 @@
 #  src/toil/lib/humanize.py (bytes2human; human2bytes)
 #  src/toil/batchSystems/__init__.py (MemoryString)
 #
+from functools import total_ordering
 
 
 def bytes_in_unit(unit: str = 'B') -> int:
@@ -35,3 +36,47 @@ def convert_units(num: float,
     assert src_unit in units, f"{src_unit} not a valid unit, valid units are {units}."
     assert dst_unit in units, f"{dst_unit} not a valid unit, valid units are {units}."
     return (num * bytes_in_unit(src_unit)) / bytes_in_unit(dst_unit)
+
+
+@total_ordering
+class MemoryString:
+    """
+    Represents an amount of bytes, as a string, using suffixes for the unit.
+
+    Comparable based on the actual number of bytes instead of string value.
+    """
+    def __init__(self, string):
+        if string[-1] == 'K' or string[-1] == 'M' or string[-1] == 'G' or string[-1] == 'T': #10K
+            self.unit = string[-1]
+            self.val = float(string[:-1])
+        elif len(string) >= 3 and (string[-2] == 'k' or string[-2] == 'M' or string[-2] == 'G' or string[-2] == 'T'):
+            self.unit = string[-2]
+            self.val = float(string[:-2])
+        else:
+            self.unit = 'B'
+            self.val = float(string)
+        self.bytes = self.byteVal()
+
+    def __str__(self):
+        if self.unit != 'B':
+            return str(self.val) + self.unit
+        else:
+            return str(self.val)
+
+    def byteVal(self):
+        if self.unit == 'B':
+            return self.val
+        elif self.unit == 'K':
+            return self.val * 1024
+        elif self.unit == 'M':
+            return self.val * 1048576
+        elif self.unit == 'G':
+            return self.val * 1073741824
+        elif self.unit == 'T':
+            return self.val * 1099511627776
+
+    def __eq__(self, other):
+        return self.bytes == other.bytes
+
+    def __lt__(self, other):
+        return self.bytes < other.bytes
